Added complex retry mechanism to submission creation.
To retry submission creation (and POST in general), we need
to be able to verify the status of the application, and see if
a submission was actually created. If it was, we delete it and
try again, because we don't know what state it's in (but it's
probably bad if we hit an error).

This required to change the withRetry function to allow the error
predicate to promise a boolean, since the recovery phase has to do
some additional web requests.