Users/kp101090/flighting (#15)
* Restructured project to factor out common code.

The API helper is now in its own "common" folder independent of
any task. The gulpfile has been extensively modified to support
this. Adding new tasks/components should be relatively easier
in the future, and will prevent code duplication.

* Further removed common code out of the publish task.

The code in question is mostly to deal with general "submissions".
It will be reused in the flight task.

* Added definitions/stubs for flight task

* Added in-depth contribution guide

* Fixed bug where getAppIdByName returned an app resource.

* Made pollSubmissionStatus use the GET /status API.

This allows us to not redownload the entire metadata
every time we poll for the status.

* Added stubs for flighting behavior

* Filled in stubs for flighting behavior (untested)

* Added UI code for Flight task (untested).

* Update zip logic and couple of URLs

* Refactoring + documentation

* Cleanup obsolete comments + fix as per PR feedback

* Update publish task to clean up input

* Make use of azure-storage module for file upload

* Add retry policy and timeout options

* Split options on multiple lines

* Add package to zip using '_' instead of '/'